backend-services/data-service/requirements.txt
Flask
requests
pymongo
yfinance
finnhub-python
python-dotenv